feat: update file storage path and add new file list components
- Changed the ROOT directory for file storage from "/mnt/storage" to "/app/static".
- Updated nginx configuration to comment out autoindex for performance optimization.
- Added `framer-motion` library for animations.
- Introduced new components: `FileItemGrid`, `FileItemList`, and `LoadingSkeleton` for improved file listing UI.
- Enhanced `FileList` component to support both grid and list views with infinite scrolling.
- Implemented file icon rendering based on file types in `utils.tsx`.
- Added utility functions for formatting file sizes and dates.
