+++ nginx.conf
@@ -10,7 +10,7 @@ server {
         alias /app/static/;
         
         # Performance optimizations
-        autoindex on;
+        # autoindex on;
         sendfile on;
         tcp_nopush on;
         aio on;
