+++ package.json
@@ -13,6 +13,7 @@
     "@monaco-editor/react": "^4.7.0",
     "@types/busboy": "^1.5.4",
     "busboy": "1.6.0",
+    "framer-motion": "^12.23.12",
     "monaco-editor": "^0.52.2",
     "next": "15.5.1",
     "next-themes": "^0.4.6",
